--- docs/todos.html
@@ -1,5 +1,0 @@
-<html><head><link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" /><script src="https://code.jquery.com/jquery-1.12.2.min.js"></script><script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script><style> pre { width: 95%; } </style></head><body><div class="container"><div class="jumbotron"><h1>Todos</h1></div><div class="panel panel-default"><div class="panel-heading" role="tab" id="heading-todos-GET"><h4 class="panel-title"><a role="button" data-toggle="collapse" data-parent="#accordion" aria-expanded="true" aria-controls="collapse-todos-GET" href="#collapse-todos-GET">GET /todos</a></h4></div><div id="collapse-todos-GET" class="panel-collapse collapse" role="tabpanel" aria-labelledby="heading-todos-GET"><div class="panel-body"><h4>Request Data</h4><div class="container"><h5>Headers</h5><pre>{}</pre></div></div><div class="panel-body"><h4>Response Data</h4><div class="container"><h5>Status: 200</h5><h5>Headers</h5><pre>{}</pre><h5>Body</h5><pre>[
-  {
-    "name": "a"
-  }
-]</pre></div></div></div></div></div></body></html>
