+++ package.json
@@ -13,17 +13,17 @@
     "lint": "tsc --noEmit && eslint \"*/**/*.ts\"",
     "lint-diff": "git diff --name-only --cached --relative | grep \\\\.ts$ | xargs eslint",
     "lint-fix": "eslint \"**/*.ts\" --fix",
-    "migrations": "ts-node-dev ./node_modules/typeorm/cli.js migration:run --config ./config/typeorm.ts",
-    "migrations-undo": "ts-node-dev ./node_modules/typeorm/cli.js migration:revert --config ./config/typeorm.ts",
-    "migrations-test": "NODE_ENV=testing ts-node-dev ./node_modules/typeorm/cli.js migration:run --config ./config/typeorm.ts",
-    "migrations-create": "ts-node-dev ./node_modules/typeorm/cli.js migration:create --config ./config/typeorm.ts -n",
-    "migrations-generate": "ts-node-dev ./node_modules/typeorm/cli.js migration:generate --config ./config/typeorm.ts -n",
+    "migrations": "ts-node-dev ./node_modules/typeorm/cli.js migration:run --config src/db/config.ts",
+    "migrations-undo": "ts-node-dev ./node_modules/typeorm/cli.js migration:revert --config src/db/config.ts",
+    "migrations-test": "NODE_ENV=testing ts-node-dev ./node_modules/typeorm/cli.js migration:run --config src/db/config.ts",
+    "migrations-create": "ts-node-dev ./node_modules/typeorm/cli.js migration:create --config src/db/config.ts -n",
+    "migrations-generate": "ts-node-dev ./node_modules/typeorm/cli.js migration:generate --config src/db/config.ts -n",
     "outdated": "npm outdated --depth 0",
     "postinstall": "npm run build",
     "prestart": "npm run lint",
     "pretest": "npm run lint",
-    "dev": "ts-node-dev --respawn --transpileOnly ./server.ts",
-    "start": "npm run build && node ./dist/server.js",
+    "dev": "ts-node-dev --respawn --transpileOnly ./src/server.ts",
+    "start": "npm run build && node ./dist/src/server.js",
     "test": "npm run test-basic -- --runInBand",
     "test-basic": "NODE_ENV=testing jest --forceExit --detectOpenHandles --verbose",
     "test-inspect": "NODE_ENV=testing node --inspect --debug-brk --detectOpenHandles jest"
@@ -31,7 +31,7 @@
   "cacheDirectories": [
     "node_modules"
   ],
-  "main": "app.js",
+  "main": "src/server.ts",
   "author": "Wolox",
   "homepage": "https://github.com/mpicollo-wolox/typescript-base",
   "license": "MIT",
