+++ src/api/controllers/todos.ts
@@ -1,5 +1,5 @@
 import { Response, NextFunction, Request } from 'express';
-import { getAllTodos, Todo } from '../services/todos';
+import { getAllTodos, Todo } from '../../services/todos';
 
 export function getTodos(_: Request, res: Response, next: NextFunction): Promise<Response> {
   return getAllTodos()
