+++ src/console.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 
 import repl from 'repl';
 
-import services from './app/services';
+import services from './services';
 
 const pjson = JSON.parse(fs.readFileSync('./package.json', 'utf-8'));
