+++ src/services/users.ts
@@ -1,7 +1,5 @@
-import { DeepPartial } from 'utils';
-
-import { getRepository, FindManyOptions, FindConditions, Repository } from 'typeorm';
-import { User } from '../models/user';
+import { getRepository, FindManyOptions, FindConditions, Repository, DeepPartial } from 'typeorm';
+import { User } from '../db/models/user';
 
 const userRepository = (): Repository<User> => getRepository(User);
