+++ tests/setup.ts
@@ -1,5 +1,5 @@
 import { getConnection, createConnection } from 'typeorm';
-import typeOrmConfig from '../config/typeorm';
+import typeOrmConfig from '../src/db/config';
 
 const truncateDatabase = (): Promise<void> => getConnection().synchronize(true);
