+++ tests/todos.spec.ts
@@ -2,7 +2,7 @@ const mockedResponse = [{ name: 'a' }];
 jest.doMock('request-promise', () => (): Promise<object> => Promise.resolve(mockedResponse));
 
 import request from 'supertest';
-import app from '../app';
+import app from '../src/app';
 
 describe('todos', () => {
   describe('/todos GET', () => {
